refactor(voice): move WS send wrappers + REGISTER to voice_ws_proto (refs #353, refs #331)
Promotes s_ws → g_voice_ws (extern in voice.h, was static in voice.c)
so voice_ws_proto.c can reach the WS handle without per-call getter
churn.

Moves voice_ws_send_text + voice_ws_send_binary +
voice_ws_send_binary_public + voice_ws_send_register out of voice.c
into voice_ws_proto.c.  s_audio_drop_count moves with them (only writer
is voice_ws_send_binary, no external readers).

CMakeLists.txt registers voice_ws_proto.c.  voice.c #includes the new
voice_ws_proto.h alongside the other voice_*.h sibling modules.

E2E verification on 192.168.1.90:
  story_smoke 14/14 in 70s (baseline 14/14 in 67s)

Wave 23 SRP-A1 closure, first slice.
